tools/openmoss/bridge/requirements.txt
requests>=2.28.0
websocket-client>=1.6.0
pyyaml>=6.0
